requirements.txt
flask
pillow
numpy
tensorflow
transformers
flask-cors
